add: summary

+++ CarApp/Entities/Models/Model.cs
@@ -9,14 +9,16 @@ namespace Entities.Models
 {
     public class Model:IEntity
     {
+        #region Properties
         public int Id { get; set; }
         public int? BrandId { get; set; }
         public string Name { get; set; }
-        public float Price { get; set; }
+        public int Price { get; set; }
         public int Production { get; set; }
         public string Color { get; set; }
         public int Mph { get; set; }
-        
-        
+        #endregion
+
+
     }
 }
